requirements.txt.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
schedule==1.2.0
pytest==7.4.3
python-dotenv==1.0.0
lxml==4.9.3